+++ src/3-three/sentences/conjugate/toNegative.js
@@ -2,4 +2,8 @@ const toNegative = function (s) {
   s.verbs().first().toNegative().compute('chunks')
   return s
 }
-export default toNegative
+const toPositive = function (s) {
+  s.verbs().first().toPositive().compute('chunks')
+  return s
+}
+export { toNegative, toPositive }